+++ ali/common/src/main/java/com/yanny/ali/compatibility/common/QuintConsumer.java
@@ -1,6 +1,0 @@
-package com.yanny.ali.compatibility.common;
-
-@FunctionalInterface
-public interface QuintConsumer<T, U, V, X, Y> {
-    void accept(T t, U u, V v, X x, Y y);
-}