+++ alicompat/fabric/src/compat/supplementaries/java/com/yanny/alicompat/compat/supplementaries/SupplementariesCompat.java
@@ -24,6 +24,7 @@ import net.mehvahdjukaar.supplementaries.common.entities.trades.StructureMapList
 import net.mehvahdjukaar.supplementaries.common.items.loot.CurseLootFunction;
 import net.mehvahdjukaar.supplementaries.common.items.loot.RandomArrowFunction;
 import net.mehvahdjukaar.supplementaries.reg.ModTags;
+import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.npc.VillagerTrades;
@@ -56,8 +57,11 @@ public class SupplementariesCompat implements IModCompat {
         registry.registerItemListing(StructureMapListing.class, SupplementariesCompat::getStructureMapListingNode);
         registry.registerItemListing(RandomAdventurerMapListing.class, SupplementariesCompat::getAdventurerMapListingNode);
 
+        ResourceLocation redMerchant = new ResourceLocation(SupplementariesLang.MOD_ID, "red_merchant");
+
         registry.registerTrades(
-                new ResourceLocation(SupplementariesLang.MOD_ID, "red_merchant"),
+                redMerchant,
+                BuiltInRegistries.ENTITY_TYPE.get(redMerchant),
                 SupplementariesCompat::getRedMerchantTrades,
                 (level) -> new TradeLevelInfo(new RangeValue(RED_MERCHANT_TRADE_COUNT))
         );